SPlit FIber JS into individual files
